% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_and_rescale.R
\name{attach_and_rescale}
\alias{attach_and_rescale}
\title{Attach and Rescale
This formula is a wrapper for the final portion of the Index Calculation. It takes calculated distances, binds it with
mangroves layer, calculates a mean of the threats and finally rescale the index in a value from 0 to 1.}
\usage{
attach_and_rescale(x)
}
\arguments{
\item{x}{is the distance calculated form the \code{calc_distance.R} function.}
}
\value{
spatial object
}
\description{
Attach and Rescale
This formula is a wrapper for the final portion of the Index Calculation. It takes calculated distances, binds it with
mangroves layer, calculates a mean of the threats and finally rescale the index in a value from 0 to 1.
}
\examples{

\dontrun{

attach_and_rescale(distance)
}

}
