% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{calc_distance}
\alias{calc_distance}
\title{calc_distance}
\usage{
calc_distance(x)
}
\arguments{
\item{x}{must be a vector file representing a threat. The formula will calculate the distance between it and mangrove centroids}
}
\value{
spatial object form sf package
}
\description{
Calculate the distance of mangrove patches from threats.
}
\examples{
\dontrun{
thr <- lapply(FileList, sf::st_as_sf)
mangrove <- sf::st_as_sf(mangrove_layer)
centroid <- sf::st_point_on_surface(mangrove)
lapply(thr, calc_distance)
 }

}
