% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{Cor2Cov}
\alias{Cor2Cov}
\title{Correlation to Covariance}
\usage{
Cor2Cov(A, v)
}
\arguments{
\item{A}{Numeric matrix.
A \eqn{p \times p} positive definite correlation matrix.}

\item{v}{Numeric vector of length \code{p} or \verb{p by 1} matrix.
A vector of \eqn{p \times 1} standard deviations.}
}
\value{
Returns a covariance matrix.
}
\description{
Converts a correlation matrix to a covariance matrix.
}
\details{
The covariance matrix is given by

\deqn{
    \left(
      \mathbf{v}
      \mathbf{I}
    \right)
    \mathbf{A}
    \left(
      \mathbf{v}
      \mathbf{I}
    \right)
  }

where
\itemize{
\item \eqn{\mathbf{A}_{p \times p}} is a correlation matrix,
\item \eqn{\mathbf{I}_{p \times p}} is an identity matrix,
\item \eqn{\mathbf{v}_{p \times 1}} is a vector of standard deviations.
}
}
\examples{
A <- matrix(
  data = c(
    1, 0.509902, 0.26,
    0.509902, 1, 0.509902,
    0.26, 0.509902, 1
  ),
  ncol = 3
)
v <- c(15, 15, 15)

diag(v) \%*\% A \%*\% diag(v)

Cor2Cov(A, v)
}
\seealso{
Other conversion functions: 
\code{\link{Cov2Cor}()},
\code{\link{low2sym}()},
\code{\link{up2sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{conversion functions}
\keyword{conversion}
