% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dot.R
\name{Dot}
\alias{Dot}
\alias{Dot.default}
\alias{Dot.yac_symbol}
\title{Dot Product of Tensors}
\usage{
Dot(t1, t2, ...)

\method{Dot}{default}(t1, t2, ...)

\method{Dot}{yac_symbol}(t1, t2, str = TRUE, ...)
}
\arguments{
\item{t1}{Tensor. Vector or matrix.}

\item{t2}{Tensor. Vector or matrix.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, the function returns a character string.
If \code{FALSE}, the function returns an \code{R} expression.}
}
\value{
Numeric.
}
\description{
Returns the dot product (or inner product) of two tensors.
}
\details{
This is a wrapper to the \code{\%*\%} infix operator and
a compatibility layer to the \code{Yacas} \code{Dot} command.

If any of the arguments is a scalar,
the function performs scalar multiplication.
}
\examples{
Dot(c(1, 2), 3)
Dot(2, c(3, 4))
Dot(c(1, 2), c(3, 4))
Dot(rbind(c(1, 2), c(3, 4)), c(5, 6))
Dot(c(5, 6), rbind(c(1, 2), c(3, 4)))
Dot(rbind(c(1, 2), c(3, 4)), rbind(c(5, 6), c(7, 8)))
}
\seealso{
Other operation functions: 
\code{\link{\%.\%}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{operation functions}
\keyword{operation}
