% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dot.R
\name{\%.\%}
\alias{\%.\%}
\title{Dot Product}
\usage{
t1 \%.\% t2
}
\arguments{
\item{t1}{Tensor. Vector or matrix.}

\item{t2}{Tensor. Vector or matrix.}
}
\value{
Numeric.
}
\description{
Returns the dot product (or inner product) of two tensors.
}
\details{
This is a wrapper to the \code{\%*\%} infix operator and
a compatibility layer to the \code{Yacas} \code{Dot} command.

If any of the arguments is a scalar,
the function performs scalar multiplication.
}
\examples{
c(1, 3, -5) \%.\% c(4, -2, -1)
}
\seealso{
Other operation functions: 
\code{\link{Dot}()},
\code{\link{Transpose}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{operation functions}
\keyword{operation}
