% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identity.R
\name{Identity}
\alias{Identity}
\title{Identity Matrix \eqn{\mathbf{I}}}
\usage{
Identity(m)
}
\arguments{
\item{m}{Integer. Size.}
}
\value{
Numeric matrix.
}
\description{
Generates an identity matrix.
}
\details{
The \eqn{m \times m} identity matrix \eqn{\mathbf{I}} is given by
\deqn{
    \mathbf{I}_{m}
    =
    \mathrm{diag} \left(1, 1, \cdots, 1 \right) .
  }
}
\examples{
Identity(3)
}
\references{
\href{https://en.wikipedia.org/wiki/Identity_matrix}{Wikipedia: Identity matrix}
}
\seealso{
Other constructor functions: 
\code{\link{IdentityFrom}()},
\code{\link{ZeroMatrixFrom}()},
\code{\link{ZeroMatrix}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{constructor functions}
\keyword{constructor}
