% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identity.R
\name{IdentityFrom}
\alias{IdentityFrom}
\title{Identity Matrix \eqn{\mathbf{I}} from Dimensions of Another Square Matrix}
\usage{
IdentityFrom(A)
}
\arguments{
\item{A}{Input.}
}
\value{
Numeric matrix.
}
\description{
Generates an identity matrix
using the dimensions of an \eqn{m \times m} matrix \eqn{\mathbf{A}}.
}
\details{
The \eqn{m \times m} identity matrix \eqn{\mathbf{I}} is given by
\deqn{
    \mathbf{I}_{m}
    =
    \mathrm{diag} \left(1, 1, \cdots, 1 \right) .
  }
}
\examples{
IdentityFrom(matrix(1:9, ncol = 3))
}
\references{
\href{https://en.wikipedia.org/wiki/Identity_matrix}{Wikipedia: Identity matrix}
}
\seealso{
Other constructor functions: 
\code{\link{Identity}()},
\code{\link{ZeroMatrixFrom}()},
\code{\link{ZeroMatrix}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{constructor functions}
\keyword{constructor}
