% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_diag}
\alias{is_diag}
\title{Is the Matrix Diagonal?}
\usage{
is_diag(A, tol = 1e-08)
}
\arguments{
\item{A}{Matrix.}

\item{tol}{Numeric.
Tolerance.}
}
\description{
Checks if off-diagonal elements of a square matrix are all zeroes.
}
\details{
The off-diagonal elements are added and compared to a tolerance value.
If the sum is less than or equal to the tolerance value,
all the elements are assumed to be zeroes.
}
\examples{
I <- diag(3)
I
is_diag(I)
}
\references{
\href{https://en.wikipedia.org/wiki/Diagonal_matrix}{Wikipedia: Diagonal matrix}
}
\seealso{
Other is functions: 
\code{\link{is_idempot}()},
\code{\link{is_inv}()},
\code{\link{is_nilpot}()},
\code{\link{is_posdef}()},
\code{\link{is_sing}()},
\code{\link{is_sqr}()},
\code{\link{is_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{is functions}
\keyword{is}
