% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_posdef}
\alias{is_posdef}
\title{Is the Matrix Positive Definite?}
\usage{
is_posdef(X, tol = 1e-08)
}
\arguments{
\item{X}{Numeric matrix.}

\item{tol}{Numeric.
Tolerance.}
}
\description{
Is the Matrix Positive Definite?
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is_posdef(X = Sigma)
}
\references{
\href{https://en.wikipedia.org/wiki/Definiteness_of_a_matrix}{Wikipedia: Definiteness of a Matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
