% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sqr}
\alias{is_sqr}
\title{Is the Matrix Square?}
\usage{
is_sqr(A, chk.num = TRUE)
}
\arguments{
\item{A}{Matrix.}

\item{chk.num}{Logical.
Check if the input matrix is numeric.}
}
\description{
Checks if the dimensions of a matrix are identical.
}
\details{
\deqn{
    \left(
      \mathbf{A}_{m \times m}
    \right)
  }
}
\examples{
A <- matrix(
  data = 1:9,
  nrow = 3
)
dim(A)
B <- matrix(
  data = 1:10,
  ncol = 2
)
dim(B)
# Returns TRUE
is_sqr(A)
# Returns FALSE
is_sqr(B)
}
\references{
\href{https://en.wikipedia.org/wiki/Square_matrix}{Wikipedia: Square matrix}
}
\seealso{
Other is functions: 
\code{\link{is_diag}()},
\code{\link{is_idempot}()},
\code{\link{is_inv}()},
\code{\link{is_nilpot}()},
\code{\link{is_posdef}()},
\code{\link{is_sing}()},
\code{\link{is_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{is functions}
\keyword{is}
