% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sym}
\alias{is_sym}
\title{Is the Matrix Symmetric?}
\usage{
is_sym(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\description{
Is the Matrix Symmetric?
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is_sym(X = Sigma)
}
\author{
Ivan Jacob Agaloos Pesigan
}
