% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsDiagonal.R
\name{IsDiagonal}
\alias{IsDiagonal}
\alias{IsDiagonal.default}
\alias{IsDiagonal.yac_symbol}
\title{Test for a Diagonal Matrix}
\usage{
IsDiagonal(A, ...)

\method{IsDiagonal}{default}(A, tol = 1e-08, ...)

\method{IsDiagonal}{yac_symbol}(A, ...)
}
\arguments{
\item{A}{Input.}

\item{...}{...}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
\code{\link[=IsDiagonal]{IsDiagonal()}} returns \code{TRUE} if \code{A} is a diagonal matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} matrix \eqn{\mathbf{A}} is diagonal
if all off-diagonal elements are zero.

\code{A} is considered to be a diagonal matrix
if the absolute value of all off-diagonal elements
is less than or equal to a tolerance value.
}
\examples{
# TRUE
IsDiagonal(diag(2))

# FALSE
IsDiagonal(matrix(1:9, ncol = 3))
}
\references{
\href{https://en.wikipedia.org/wiki/Diagonal_matrix}{Wikipedia: Diagonal matrix}
}
\seealso{
Other predicate functions: 
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
