% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsIdempotent.R
\name{IsIdempotent}
\alias{IsIdempotent}
\alias{IsIdempotent.default}
\alias{IsIdempotent.yac_symbol}
\title{Test for an Idempotent Matrix}
\usage{
IsIdempotent(A, ...)

\method{IsIdempotent}{default}(A, tol = 1e-08, ...)

\method{IsIdempotent}{yac_symbol}(A, ...)
}
\arguments{
\item{A}{Input.}

\item{...}{...}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
\code{\link[=IsIdempotent]{IsIdempotent()}} returns \code{TRUE} if \code{A} is an idempotent matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} matrix \eqn{\mathbf{A}} is idempotent if
\deqn{
    \mathbf{A}^2 = \mathbf{A} .
  }

\code{A} is considered to be an idempotent matrix
if \code{all(abs(A \%*\% A - A) <= tol)} returns \code{TRUE}.
}
\examples{
# TRUE
IsIdempotent(matrix(data = c(1, 0, 0, 1), ncol = 2))

# FALSE
IsIdempotent(matrix(data = 1, nrow = 2, ncol = 2))
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
