% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsInvertible.R
\name{IsInvertible}
\alias{IsInvertible}
\title{Test for an Invertible Matrix}
\usage{
IsInvertible(A, tol = 1e-08)
}
\arguments{
\item{A}{Input.}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
\code{\link[=IsInvertible]{IsInvertible()}} returns \code{TRUE} if \code{A} is an invertible matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} matrix \eqn{\mathbf{A}} is invertible
if there exists an \eqn{m \times m} matrix \eqn{\mathbf{B}}
such that
\deqn{
    \mathbf{A} \mathbf{B}
    =
    \mathbf{B}
    \mathbf{A}
    =
    \mathbf{I}_{m}
  }
where \eqn{\mathbf{I}_{m}} is an \eqn{m \times m} identity matrix.

An \eqn{m \times m} matrix \eqn{\mathbf{A}} is noninvertible
or singular if its determinant is zero.

\code{A} is considered invertible if its determinant
is \strong{NOT} less than or equal
to a tolerance value.
}
\examples{
# TRUE
IsInvertible(diag(2))

# FALSE
IsInvertible(matrix(1, nrow = 2, ncol = 2))
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
