% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsMatrix.R
\name{IsMatrix}
\alias{IsMatrix}
\alias{IsMatrix.default}
\alias{IsMatrix.yac_symbol}
\title{Test for a Matrix}
\usage{
IsMatrix(A)

\method{IsMatrix}{default}(A)

\method{IsMatrix}{yac_symbol}(A)
}
\arguments{
\item{A}{Input.}
}
\value{
Logical.
}
\description{
\code{\link[=IsMatrix]{IsMatrix()}} returns \code{TRUE} if \code{A} is a matrix,
and \code{FALSE} otherwise.
}
\details{
A matrix is a rectangular array of numbers, symbols, or expressions,
with \eqn{m} rows by \eqn{n} columns
typically denoted by uppercase roman letters such as
\eqn{\mathbf{A}}, \eqn{\mathbf{B}}, and \eqn{\mathbf{C}}.

\deqn{
    \mathbf{A}
    =
    \begin{pmatrix}
    	a_{1, 1} & a_{1, 2} & a_{1, 3} & \cdots & a_{1, n} \\
    	a_{2, 1} & a_{2, 2} & a_{2, 3} & \cdots & a_{2, n} \\
    	\vdots   & \vdots   & \vdots   & \cdots & \vdots   \\
    	a_{i, j} & a_{i, j} & a_{i, j} & \cdots & a_{i, j} \\
    	\vdots   & \vdots   & \vdots   & \vdots & \vdots   \\
    	a_{m, 1} & a_{m, 2} & a_{m, 3} & \cdots & a_{m, n}
    \end{pmatrix}
    =
    \left( a_{i, j} \in \mathbf{R}^{m \times n} \right)
  }
}
\examples{
# TRUE
IsMatrix(matrix(1:10, ncol = 2))
IsMatrix(matrix(1:10, nrow = 2))

# FALSE
IsMatrix(1:5)
}
\references{
\href{https://en.wikipedia.org/wiki/Matrix_(mathematics)}{Wikipedia: Matrix (mathematics)}
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsNilpotent}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
