% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsNilpotent.R
\name{IsNilpotent}
\alias{IsNilpotent}
\title{Test for a Nilpotent Matrix}
\usage{
IsNilpotent(A, tol = 1e-08)
}
\arguments{
\item{A}{Input.}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
\code{\link[=IsNilpotent]{IsNilpotent()}} returns \code{TRUE} if \code{A} is a nilpotent matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} matrix \eqn{\mathbf{A}} is nilpotent
if all the eigenvalues are zero.

\code{A} is considered to be a nilpotent matrix
if the absolute value of all eigenvalues
is less than or equal to a tolerance value.
}
\examples{
# TRUE
IsNilpotent(matrix(data = c(0, 0, 1, 0), ncol = 2))

# FALSE
IsNilpotent(matrix(data = c(1, 0, 0, 0), ncol = 2))
}
\references{
\href{https://en.wikipedia.org/wiki/Nilpotent_matrix}{Wikipedia: Nilpotent matrix}
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
