% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsOrthogonal.R
\name{IsOrthogonal}
\alias{IsOrthogonal}
\alias{IsOrthogonal.default}
\alias{IsOrthogonal.yac_symbol}
\title{Test for an Orthogonal Matrix}
\usage{
IsOrthogonal(A, ...)

\method{IsOrthogonal}{default}(A, tol = 1e-08, ...)

\method{IsOrthogonal}{yac_symbol}(A, ...)
}
\arguments{
\item{A}{Input.}

\item{...}{...}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
\code{\link[=IsOrthogonal]{IsOrthogonal()}} returns \code{TRUE} if \code{A} is an orthogonal matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} matrix \eqn{\mathbf{A}} is orthogonal if
\deqn{
    \mathbf{A}^{\mathsf{T}} \mathbf{A}
    =
    \mathbf{A} \mathbf{A}^{\mathsf{T}}
    \mathbf{I} .
  }

Or equivalently
\deqn{
    \mathbf{A}^{\mathsf{T}}
    =
    \mathbf{A}^{-1} .
  }

\code{A} is considered to be an orthogonal matrix
if \code{all(abs(crossprod(A) - diag(dim(A)[1])) <= tol)} returns \code{TRUE}.
}
\examples{
# TRUE
IsOrthogonal(matrix(data = c(1, 0, 0, 1), ncol = 2))

# FALSE
IsOrthogonal(matrix(data = 1, nrow = 2, ncol = 2))
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()},
\code{\link{MatrixCheck}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
