% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsPositiveDefinite.R
\name{IsPositiveDefinite}
\alias{IsPositiveDefinite}
\title{Test for a Positive Definite Matrix}
\usage{
IsPositiveDefinite(A, tol = 1e-08)
}
\arguments{
\item{A}{Input.}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
\code{\link[=IsPositiveDefinite]{IsPositiveDefinite()}} returns \code{TRUE} if \code{A}
is a positive definite matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} symmetric matrix \eqn{\mathbf{A}}
is positive definite
if all of its eigenvalues are positive.

\code{A} is considered to be a positive definite matrix
if \strong{NONE} of its eigenvalues are less than or equal to
a tolerance value.
}
\examples{
# TRUE
IsPositiveDefinite(diag(2))

# FALSE
IsPositiveDefinite(matrix(1:9, ncol = 3))
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
