% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsScalar.R
\name{IsScalar}
\alias{IsScalar}
\alias{IsScalar.default}
\alias{IsScalar.yac_symbol}
\title{Test for a Scalar}
\usage{
IsScalar(k)

\method{IsScalar}{default}(k)

\method{IsScalar}{yac_symbol}(k)
}
\arguments{
\item{k}{Input.}
}
\value{
Logical.
}
\description{
\code{\link[=IsScalar]{IsScalar()}} returns \code{TRUE} if \code{k} is a scalar,
and \code{FALSE} otherwise.
}
\details{
A scalar is any real number that can be measured
using a single real number.
A scalar is typically denoted by \eqn{k}.

\code{k} is considered to be a scalar
if it can be coerced to a vector
and \code{length(k) == 1} returns \code{TRUE}.
}
\examples{
# TRUE
IsScalar(1)
IsScalar(as.matrix(1))

# FALSE
IsScalar(c(1, 2, 3))
IsScalar(as.matrix(c(1, 2, 3)))
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
