% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsSingular.R
\name{IsSingular}
\alias{IsSingular}
\title{Test for a Singular Matrix}
\usage{
IsSingular(A, tol = 1e-08)
}
\arguments{
\item{A}{Input.}

\item{tol}{Numeric. Tolerance.}
}
\value{
Logical.
}
\description{
\code{\link[=IsInvertible]{IsInvertible()}} returns \code{TRUE} if \code{A} is a singular matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} matrix \eqn{\mathbf{A}} is singular
if its determinant is zero.

The function compares the determinant of \eqn{\mathbf{A}}
to a tolerance value.
If the determinant is less than or equal
to a tolerance value,
the matrix is assumed to be singular.
}
\examples{
# TRUE
IsSingular(matrix(1, nrow = 2, ncol = 2))

# FALSE
IsSingular(diag(2))
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
