% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsSquareMatrix.R
\name{IsSquareMatrix}
\alias{IsSquareMatrix}
\alias{IsSquareMatrix.default}
\alias{IsSquareMatrix.yac_symbol}
\title{Test for a Square Matrix}
\usage{
IsSquareMatrix(A)

\method{IsSquareMatrix}{default}(A)

\method{IsSquareMatrix}{yac_symbol}(A)
}
\arguments{
\item{A}{Input.}
}
\value{
Logical.
}
\description{
\code{\link[=IsSquareMatrix]{IsSquareMatrix()}} returns \code{TRUE} if \code{A} is a square matrix,
and \code{FALSE} otherwise.
}
\details{
A matrix \eqn{\mathbf{A}} is square
if it has the same number of rows and columns
\deqn{
    \mathbf{A}_{m \times m} .
  }

\code{A} is considered to be a square matrix
if \code{dim(A)[1] == dim(A)[2]} returns \code{TRUE}.
}
\examples{
# TRUE
IsSquareMatrix(matrix(1:9, ncol = 3))

# FALSE
IsSquareMatrix(1:5)
IsSquareMatrix(matrix(1:10, nrow = 2))
}
\references{
\href{https://en.wikipedia.org/wiki/Square_matrix}{Wikipedia: Square matrix}
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()},
\code{\link{MatrixCheck}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
