% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsSymmetric.R
\name{IsSymmetric}
\alias{IsSymmetric}
\alias{IsSymmetric.default}
\alias{IsSymmetric.yac_symbol}
\title{Test for a Symmetric Matrix}
\usage{
IsSymmetric(A)

\method{IsSymmetric}{default}(A)

\method{IsSymmetric}{yac_symbol}(A)
}
\arguments{
\item{A}{Input.}
}
\value{
Logical.
}
\description{
\code{\link[=IsSymmetric]{IsSymmetric()}} returns \code{TRUE} if \code{A} is a symmetric matrix,
and \code{FALSE} otherwise.
}
\details{
An \eqn{m \times m} matrix \eqn{\mathbf{A}} is symmetric if
\deqn{
    \mathbf{A}
    =
    \mathbf{A}^{\mathsf{T}} .
  }

\code{A} is considered to be a symmetric matrix
if \code{all(A == t(A))} returns \code{TRUE}.
}
\examples{
# TRUE
A <- matrix(
  data = c(1, 2, 3, 2, 4, 5, 3, 5, 6),
  ncol = 3
)
IsSymmetric(A)

# FALSE
IsSymmetric(matrix(1:10, ncol = 2))
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsVector}()},
\code{\link{MatrixCheck}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
