% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsVector.R
\name{IsVector}
\alias{IsVector}
\alias{IsVector.default}
\alias{IsVector.yac_symbol}
\title{Test for a Vector}
\usage{
IsVector(v)

\method{IsVector}{default}(v)

\method{IsVector}{yac_symbol}(v)
}
\arguments{
\item{v}{Input.}
}
\value{
Logical.
}
\description{
\code{\link[=IsVector]{IsVector()}} returns \code{TRUE} if \code{v} is a vector,
and \code{FALSE} otherwise.
}
\details{
A vector is an array of numbers, symbols, or expressions,
typically denoted by lowercase roman letters such as
\eqn{\mathbf{v}} and \eqn{\mathbf{u}}.

\code{v} is considered to be a vector
if \code{is.vector(v)} returns \code{TRUE} or
\code{v} is a \code{matrix}
and \code{dim(v)[1] == 1 | dim(v)[2] == 1} returns \code{TRUE}.
}
\examples{
# TRUE
IsVector(1:5)
IsVector(matrix(1:5, ncol = 1))
IsVector(matrix(1:5, nrow = 1))

# FALSE
IsVector(matrix(1:10, ncol = 2))
}
\references{
\href{https://en.wikipedia.org/wiki/Row_and_column_vectors}{Wikipedia: Row and column vectors}
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{MatrixCheck}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
