% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{LS}
\alias{LS}
\title{Least Squares}
\usage{
LS(A, b)
}
\arguments{
\item{A}{\verb{m by n} numeric matrix.}

\item{b}{Numeric vector of length \code{m} or \verb{m by 1} matrix.}
}
\description{
Least squares solution
for \eqn{\mathbf{A} \mathbf{x} = \mathbf{b}}
}
\details{
The least squares solution in given by
\deqn{
    \mathbf{A}
    \mathbf{x}
    =
    \mathbf{b} \\
    \mathbf{A}^{\mathsf{T}}
    \mathbf{A}
    \mathbf{x}
    =
    \mathbf{A}^{\mathsf{T}}
    \mathbf{b} \\
    \mathbf{x}
    =
    \left(
      \mathbf{A}^{\mathsf{T}}
      \mathbf{A}
    \right)^{-1}
    \mathbf{A}^{\mathsf{T}}
    \mathbf{b}
  }
}
\examples{
X <- MASS::mvrnorm(
  n = 100,
  mu = c(0, 0),
  Sigma = matrix(
    data = c(1, .5, .5, 1),
    ncol = 2
  ),
)
A <- cbind(1, X[, 1])
b <- X[, 2]
LS(A, b)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{solve functions}
\keyword{solve}
