% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixCheck.R
\name{MatrixCheck}
\alias{MatrixCheck}
\alias{MatrixCheck.default}
\alias{MatrixCheck.yac_symbol}
\title{Check Matrix Properties}
\usage{
MatrixCheck(
  A,
  IsSquareMatrix = FALSE,
  IsSymmetric = FALSE,
  IsNilpotent = FALSE,
  ...
)

\method{MatrixCheck}{default}(
  A,
  IsSquareMatrix = FALSE,
  IsSymmetric = FALSE,
  IsNilpotent = FALSE,
  ...
)

\method{MatrixCheck}{yac_symbol}(
  A,
  IsSquareMatrix = FALSE,
  IsSymmetric = FALSE,
  IsNilpotent = FALSE,
  ...
)
}
\arguments{
\item{A}{Input.}

\item{IsSquareMatrix}{Logical.
Test if input is a square matrix.}

\item{IsSymmetric}{Logical.
Test if input is a symmetric matrix.}

\item{IsNilpotent}{Logical.
Test if input is a nilponent matrix.}

\item{...}{...}
}
\description{
Checks matrix properties and stops if any test fails.
Returns the matrix if all tests succeed.
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsMatrix}()},
\code{\link{IsNilpotent}()},
\code{\link{IsOrthogonal}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
