% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixR.R
\docType{package}
\name{matrixR}
\alias{matrixR}
\title{matrixR: Matrix/Linear Algebra}
\description{
A collection of functions related to Matrix/Linear Algebra.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{description}
