% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesandzeroes.R
\name{ones}
\alias{ones}
\title{Identity Matrix}
\usage{
ones(n)
}
\arguments{
\item{n}{Integer. Size.}
}
\description{
Generates an indentity matrix.
}
\examples{
ones(3)
}
\references{
\href{https://en.wikipedia.org/wiki/Identity_matrix}{Wikipedia: Identity matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
