% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesandzeroes.R
\name{ones_from}
\alias{ones_from}
\title{Identity Matrix from Dimensions of Another Square Matrix}
\usage{
ones_from(A)
}
\arguments{
\item{A}{Matrix.}
}
\description{
Generates an indentity matrix using the dimensions of another square matrix.
}
\examples{
A <- matrix(
  data = 1:9,
  ncol = 3
)
ones_from(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Identity_matrix}{Wikipedia: Identity matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
