% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Matrix Power}
\usage{
pow(A, n)
}
\arguments{
\item{A}{Matrix.}

\item{n}{Integer. Power.}
}
\description{
Calculates the power of a matrix.
}
\details{
\deqn{
    \mathbf{A}^{n}
  }
}
\examples{
A <- matrix(
  data = c(
    3, 4, 2, 5
  ),
  ncol = 2
)
pow(A, 2)
}
\author{
Ivan Jacob Agaloos Pesigan
}
