% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{tr}
\alias{tr}
\title{Matrix Trace}
\usage{
tr(A)
}
\arguments{
\item{A}{Matrix.}
}
\value{
Returns the the trace of a matrix.
}
\description{
Calculates the trace of a matrix.
}
\details{
The trace of an \eqn{n \times n} matrix \eqn{\mathbf{A}}
is given by

\deqn{
    \mathrm{Tr} \left( A \right)
    =
    \sum_{i = 1}^{n} a_{ii}
    =
    a_{11} + a_{22} + \cdots + a_{nn}
  }
}
\examples{
A <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
tr(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Trace_(linear_algebra)}{Wikipedia: Square matrix}
}
\seealso{
Other operation functions: 
\code{\link{pow}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{operation functions}
\keyword{operation}
