% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transpose.R
\name{Transpose}
\alias{Transpose}
\alias{Transpose.default}
\alias{Transpose.yac_symbol}
\title{Transpose of a Matrix}
\usage{
Transpose(A, ...)

\method{Transpose}{default}(A, ...)

\method{Transpose}{yac_symbol}(A, str = TRUE, ...)
}
\arguments{
\item{A}{Input.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, the function returns a character string.
If \code{FALSE}, the function returns an \code{R} expression.}
}
\description{
Derives the tranpose of a matrix.
}
\details{
The transpose of an \eqn{m \times n} matrix \eqn{\mathbf{A}}
is given by
\deqn{
    \mathbf{A}_{i, j}^{\mathsf{T}} = \mathbf{A}_{j, i}
  }

This is a wrapper around the \code{t} function
and a compatibility layer to the \code{Yacas} \code{Transpose} command.
}
\examples{
Transpose(matrix(data = c(1, 3, 2, 4), ncol = 2))
}
\references{
\href{https://en.wikipedia.org/wiki/Transpose}{Wikipedia: Transpose}
}
\seealso{
Other operation functions: 
\code{\link{Dot}()},
\code{\link{\%.\%}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{operation functions}
\keyword{operation}
