% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{up2sym}
\alias{up2sym}
\title{Upper Triangle to Symmetric}
\usage{
up2sym(A, chk.num = TRUE)
}
\arguments{
\item{A}{Matrix.}

\item{chk.num}{Logical.
Check if the input matrix is numeric.}
}
\value{
Returns a symmetric matrix.
}
\description{
Creates a symmetric matrix from the upper triangle
of a square matrix.
}
\examples{
A <- matrix(NA, ncol = 4, nrow = 4)
A[upper.tri(A, diag = TRUE)] <- 1:10
A
B <- up2sym(A)
B
is_sym(B)
}
\seealso{
Other conversion functions: 
\code{\link{Cor2Cov}()},
\code{\link{Cov2Cor}()},
\code{\link{low2sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{conversion functions}
\keyword{conversion}
