% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onesandzeroes.R
\name{zeroes}
\alias{zeroes}
\title{Zero Matrix}
\usage{
zeroes(m, n)
}
\arguments{
\item{m}{Integer. Number of rows.}

\item{n}{Integer. Number of columns.}
}
\description{
Generates a zero matrix.
}
\examples{
zeroes(3, 5)
}
\references{
\href{https://en.wikipedia.org/wiki/Zero_matrix}{Wikipedia: Zero matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
