% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroMatrix.R
\name{ZeroMatrix}
\alias{ZeroMatrix}
\title{Zero Matrix \eqn{\mathbf{0}}}
\usage{
ZeroMatrix(m, n = NULL)
}
\arguments{
\item{m}{Integer. Number of rows.}

\item{n}{Integer. Number of columns.
If \code{NULL}, assumed \code{m = n} is assumed.}
}
\value{
Numeric matrix.
}
\description{
Generates a zero matrix.
}
\details{
The \eqn{m \times n} zero matrix \eqn{\mathbf{0}}
is a matrix completely filled with zeroes.
}
\examples{
ZeroMatrix(3)
ZeroMatrix(3, 5)
}
\references{
\href{https://en.wikipedia.org/wiki/Zero_matrix}{Wikipedia: Zero matrix}
}
\seealso{
Other constructor functions: 
\code{\link{IdentityFrom}()},
\code{\link{Identity}()},
\code{\link{ZeroMatrixFrom}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{constructor functions}
\keyword{constructor}
