% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroMatrix.R
\name{ZeroMatrixFrom}
\alias{ZeroMatrixFrom}
\title{Zero Matrix \eqn{\mathbf{0}} from Dimensions of Another Matrix}
\usage{
ZeroMatrixFrom(A)
}
\arguments{
\item{A}{Input.}
}
\value{
Numeric matrix.
}
\description{
Generates a zero matrix
using the dimensions of an \eqn{m \times n} matrix \eqn{\mathbf{A}}.
}
\details{
The \eqn{m \times n} zero matrix \eqn{\mathbf{0}}
is a matrix completely filled with zeroes.
}
\examples{
ZeroMatrixFrom(matrix(1:9, ncol = 3))
}
\references{
\href{https://en.wikipedia.org/wiki/Zero_matrix}{Wikipedia: Zero matrix}
}
\seealso{
Other constructor functions: 
\code{\link{IdentityFrom}()},
\code{\link{Identity}()},
\code{\link{ZeroMatrix}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{constructor functions}
\keyword{constructor}
