% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{Cov2Cor}
\alias{Cov2Cor}
\title{Covariance to Correlation}
\usage{
Cov2Cor(A)
}
\arguments{
\item{A}{Numeric matrix.
A \eqn{p \times p} positive definite covariance matrix.}
}
\value{
Returns a correlation matrix.
}
\description{
Converts a covariance matrix to a correlation matrix.
}
\details{
The correlation matrix is given by

\deqn{
    \left(
      \mathbf{v}
      \mathbf{I}
    \right)^{-1}
    \mathbf{A}
    \left(
      \mathbf{v}
      \mathbf{I}
    \right)^{-1}
  }

where
\itemize{
\item \eqn{\mathbf{A}_{p \times p}} is a covariance matrix,
\item \eqn{\mathbf{I}_{p \times p}} is an identity matrix,
\item \eqn{\mathbf{v}_{p \times 1}} is a vector of standard deviations,
that is, the square root of the diagonal elements of \eqn{\mathbf{A}}.
}
}
\examples{
A <- matrix(
  data = c(
    225.0000, 114.7279, 58.5000,
    114.7279, 225.0000, 114.7279,
    58.5000, 114.7279, 225.0000
  ),
  ncol = 3
)

cov2cor(A)

Cov2Cor(A)
}
\seealso{
Other conversion functions: 
\code{\link{Cor2Cov}()},
\code{\link{low2sym}()},
\code{\link{up2sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{conversion functions}
\keyword{conversion}
