% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dot.R
\name{Dot}
\alias{Dot}
\alias{Dot.default}
\alias{Dot.yac_symbol}
\title{Dot Product}
\usage{
Dot(u, v, ...)

\method{Dot}{default}(u, v, ...)

\method{Dot}{yac_symbol}(u, v, str = TRUE, ...)
}
\arguments{
\item{u}{Numeric vector of length \code{m}.}

\item{v}{Numeric vector of length \code{m}.}

\item{...}{...}

\item{str}{Logical.
If \code{TRUE}, the function returns a character string.
If \code{FALSE}, the function returns an \code{R} expression.}
}
\value{
Numeric.
}
\description{
Returns the dot product of two vectors.
}
\details{
The dot product is the sum of the products of corresponding entries
of two vectors with the same length.
}
\examples{
Dot(c(1, 3, -5), c(4, -2, -1))
}
\references{
\href{https://en.wikipedia.org/wiki/Dot_product}{Wikipedia: Dot product}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{operation functions}
\keyword{operation}
