% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dot.R
\name{\%.\%}
\alias{\%.\%}
\title{Dot Product}
\usage{
u \%.\% v
}
\arguments{
\item{u}{Numeric vector of length \code{m}.}

\item{v}{Numeric vector of length \code{m}.}
}
\value{
Numeric.
}
\description{
Returns the dot product of two vectors.
}
\details{
The dot product is the sum of the products of corresponding entries
of two vectors with the same length.
}
\examples{
c(1, 3, -5) \%.\% c(4, -2, -1)
}
\references{
\href{https://en.wikipedia.org/wiki/Dot_product}{Wikipedia: Dot product}
}
\seealso{
Other operation functions: 
\code{\link{Dot}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{operation functions}
\keyword{operation}
