% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_idempot}
\alias{is_idempot}
\title{Is the Matrix Idempotent?}
\usage{
is_idempot(A, tol = 1e-08)
}
\arguments{
\item{A}{Matrix.}

\item{tol}{Numeric.
Tolerance.}
}
\description{
Checks if the input matrix is idempotent.
}
\details{
The square matrix \eqn{\mathbf{A}} is idempotent
if \eqn{\mathbf{A}^2 = \mathbf{A}}
}
\examples{
A <- matrix(
  data = c(
    1, 0, 0, 1
  ),
  ncol = 2
)
is_idempot(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Idempotent_matrix}{Wikipedia: Nilpotent matrix}
}
\seealso{
Other is functions: 
\code{\link{is_diag}()},
\code{\link{is_inv}()},
\code{\link{is_nilpot}()},
\code{\link{is_posdef}()},
\code{\link{is_sing}()},
\code{\link{is_sqr}()},
\code{\link{is_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{is functions}
\keyword{is}
