% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_inv}
\alias{is_inv}
\title{Is the Matrix Invertible?}
\usage{
is_inv(A, tol = 1e-08)
}
\arguments{
\item{A}{Matrix.}

\item{tol}{Numeric.
Tolerance.}
}
\description{
Checks if a square matrix is invertible.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is_inv(Sigma)
}
\references{
\href{https://en.wikipedia.org/wiki/Invertible_matrix}{Wikipedia: Invertible Matrix}

\href{https://en.wikipedia.org/wiki/Singular_matrix}{Wikipedia: Singular Matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
