% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_nilpot}
\alias{is_nilpot}
\title{Is the Matrix Nilpotent?}
\usage{
is_nilpot(A, tol = 1e-08)
}
\arguments{
\item{A}{Matrix.}

\item{tol}{Numeric.
Tolerance.}
}
\description{
Checks if the input matrix is nilpotent.
}
\details{
The square matrix \eqn{\mathbf{A}} is nilpotent
if all the eigenvalues of \eqn{\mathbf{A}} is zero.
}
\examples{
A <- zeroes(3, 3)
A[1, ] <- c(0, 1, 1)
is_nilpot(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Nilpotent_matrix}{Wikipedia: Nilpotent matrix}
}
\seealso{
Other is functions: 
\code{\link{is_diag}()},
\code{\link{is_idempot}()},
\code{\link{is_inv}()},
\code{\link{is_posdef}()},
\code{\link{is_sing}()},
\code{\link{is_sqr}()},
\code{\link{is_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{is functions}
\keyword{is}
