% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_posdef}
\alias{is_posdef}
\title{Is the Matrix Positive Definite?}
\usage{
is_posdef(A, tol = 1e-08)
}
\arguments{
\item{A}{Matrix.}

\item{tol}{Numeric.
Tolerance.}
}
\description{
Checks if eigenvalues in a square matrix are positive.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is_posdef(Sigma)
}
\references{
\href{https://en.wikipedia.org/wiki/Definiteness_of_a_matrix}{Wikipedia: Definiteness of a Matrix}
}
\seealso{
Other is functions: 
\code{\link{is_diag}()},
\code{\link{is_idempot}()},
\code{\link{is_inv}()},
\code{\link{is_nilpot}()},
\code{\link{is_sing}()},
\code{\link{is_sqr}()},
\code{\link{is_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{is functions}
\keyword{is}
