% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sing}
\alias{is_sing}
\title{Is the Matrix Singular?}
\usage{
is_sing(A, tol = 1e-08)
}
\arguments{
\item{A}{Matrix.}

\item{tol}{Numeric.
Tolerance.}
}
\description{
Checks if a square matrix is noninvertible or singular.
}
\details{
An \eqn{n \times n} matrix \eqn{\mathbf{A}} is singular
if its determinant is zero.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is_sing(Sigma)
}
\references{
\href{https://en.wikipedia.org/wiki/Invertible_matrix}{Wikipedia: Invertible Matrix}

\href{https://en.wikipedia.org/wiki/Singular_matrix}{Wikipedia: Singular Matrix}
}
\seealso{
Other is functions: 
\code{\link{is_diag}()},
\code{\link{is_idempot}()},
\code{\link{is_inv}()},
\code{\link{is_nilpot}()},
\code{\link{is_posdef}()},
\code{\link{is_sqr}()},
\code{\link{is_sym}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{is functions}
\keyword{is}
