% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sqr}
\alias{is_sqr}
\title{Is the Matrix Square?}
\usage{
is_sqr(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\description{
Is the Matrix Square?
}
\examples{
A <- matrix(
  data = 1:9,
  nrow = 3
)
dim(A)
B <- matrix(
  data = 1:10,
  ncol = 2
)
dim(B)
# Returns TRUE
is_sqr(X = B)
# Returns FALSE
is_sqr(X = A)
}
\references{
\href{https://en.wikipedia.org/wiki/Square_matrix}{Wikipedia: Square matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
