% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sqr}
\alias{is_sqr}
\title{Is the Matrix Square?}
\usage{
is_sqr(A, chk.num = TRUE)
}
\arguments{
\item{A}{Numeric matrix.}

\item{chk.num}{Logical.
Check if the input matrix is numeric.}
}
\description{
Checks if a matrix is square \eqn{\left( \mathbf{A}_{m \times m} \right)} .
}
\examples{
A <- matrix(
  data = 1:9,
  nrow = 3
)
dim(A)
B <- matrix(
  data = 1:10,
  ncol = 2
)
dim(B)
# Returns TRUE
is_sqr(B)
# Returns FALSE
is_sqr(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Square_matrix}{Wikipedia: Square matrix}
}
\author{
Ivan Jacob Agaloos Pesigan
}
