% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sym}
\alias{is_sym}
\title{Is the Matrix Symmetric?}
\usage{
is_sym(A, chk.num = TRUE)
}
\arguments{
\item{A}{Numeric matrix.}

\item{chk.num}{Logical.
Check if the input matrix is numeric.}
}
\description{
Checks if a matrix is symmetric \eqn{\left( \mathbf{A} = \mathbf{A}^{\mathsf{T}} \right)}.
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is_sym(Sigma)
}
\author{
Ivan Jacob Agaloos Pesigan
}
