% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_sym}
\alias{is_sym}
\title{Is the Matrix Symmetric?}
\usage{
is_sym(A, chk.num = TRUE)
}
\arguments{
\item{A}{Matrix.}

\item{chk.num}{Logical.
Check if the input matrix is numeric.}
}
\description{
Checks if the transpose of a matrix is identical to the original matrix.
}
\details{
\deqn{
    \left(
      \mathbf{A}
      =
      \mathbf{A}^{\mathsf{T}}
    \right)
  }
}
\examples{
Sigma <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
is_sym(Sigma)
}
\seealso{
Other is functions: 
\code{\link{is_diag}()},
\code{\link{is_idempot}()},
\code{\link{is_inv}()},
\code{\link{is_nilpot}()},
\code{\link{is_posdef}()},
\code{\link{is_sing}()},
\code{\link{is_sqr}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{is functions}
\keyword{is}
