% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsMatrix.R
\name{IsMatrix}
\alias{IsMatrix}
\alias{IsMatrix.default}
\alias{IsMatrix.yac_symbol}
\title{Test for a Matrix}
\usage{
IsMatrix(A)

\method{IsMatrix}{default}(A)

\method{IsMatrix}{yac_symbol}(A)
}
\arguments{
\item{A}{Input.}
}
\value{
Logical.
}
\description{
\code{\link[=IsMatrix]{IsMatrix()}} returns \code{TRUE} if \code{A} is a matrix,
and \code{FALSE} otherwise.
}
\details{
A matrix is a rectangular array of numbers, symbols, or expressions,
with \eqn{m} rows by \eqn{n} columns
typically denoted by uppercase roman letters such as
\eqn{\mathbf{A}}, \eqn{\mathbf{B}}, and \eqn{\mathbf{C}}.
}
\examples{
# TRUE
IsMatrix(matrix(1:10, ncol = 2))
IsMatrix(matrix(1:10, nrow = 2))

# FALSE
IsMatrix(1:5)
}
\references{
\href{https://en.wikipedia.org/wiki/Matrix_(mathematics)}{Wikipedia: Matrix (mathematics)}
}
\seealso{
Other predicate functions: 
\code{\link{IsDiagonal}()},
\code{\link{IsIdempotent}()},
\code{\link{IsInvertible}()},
\code{\link{IsNilpotent}()},
\code{\link{IsPositiveDefinite}()},
\code{\link{IsScalar}()},
\code{\link{IsSingular}()},
\code{\link{IsSquareMatrix}()},
\code{\link{IsSymmetric}()},
\code{\link{IsVector}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{predicate functions}
\keyword{predicate}
