% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{low2sym}
\alias{low2sym}
\title{Lower Triangle to Symmetric}
\usage{
low2sym(A, chk.num = TRUE)
}
\arguments{
\item{A}{Numeric matrix.}

\item{chk.num}{Logical.
Check if the input matrix is numeric.}
}
\value{
Returns a symmetric matrix.
}
\description{
Creates a symmetric matrix from the lower triangle
of a square matrix.
}
\examples{
A <- matrix(NA, ncol = 4, nrow = 4)
A[lower.tri(A, diag = TRUE)] <- 1:10
A
B <- low2sym(A)
is_sym(B)
}
\author{
Ivan Jacob Agaloos Pesigan

Ivan Jacob Agaloos Pesigan
}
