% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{pow}
\alias{pow}
\title{Matrix Power}
\usage{
pow(A, n)
}
\arguments{
\item{A}{Matrix.}

\item{n}{Integer. Power.}
}
\description{
Calculates the power of a square matrix.
}
\details{
The power of a matrix is given by

\deqn{
    \mathbf{A}^{n}
  }

Let \eqn{n} be a 0

\deqn{
    \mathbf{A}^{0}
    =
    \mathbf{A} .
  }

Let \eqn{n} be a positive integer for example \eqn{2}

\deqn{
    \mathbf{A}^{2}
    =
    \mathbf{A} \cdot \mathbf{A} .
  }

Let \eqn{n} be a negative integer for example \eqn{-2}

\deqn{
    \mathbf{A}^{-2}
    =
    \left( \mathbf{A}^{-1} \right) \cdot \left( \mathbf{A}^{-1} \right) .
  }
}
\examples{
A <- matrix(
  data = c(
    225, 112.50, 56.25,
    112.5, 225, 112.5,
    56.25, 112.50, 225
  ),
  ncol = 3
)
pow(A, 2)
}
\seealso{
Other operation functions: 
\code{\link{tr}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{operation functions}
\keyword{operation}
