% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{up2sym}
\alias{up2sym}
\title{Upper Triangle to Symmetric}
\usage{
up2sym(A)
}
\arguments{
\item{A}{Numeric matrix.}
}
\value{
Returns a symmetric matrix.
}
\description{
Creates a symmetric matrix from the upper triangle
of a square matrix.
}
\examples{
A <- matrix(NA, ncol = 4, nrow = 4)
A[upper.tri(A, diag = TRUE)] <- 1:10
A
up2sym(A)
}
\author{
Ivan Jacob Agaloos Pesigan

Ivan Jacob Agaloos Pesigan
}
