% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroes.R
\name{zeroes_from}
\alias{zeroes_from}
\title{Zero Matrix from Dimensions of Another Matrix}
\usage{
zeroes_from(A)
}
\arguments{
\item{A}{Matrix.}
}
\description{
Generates a zero matrix using the dimensions of another matrix.
}
\examples{
A <- matrix(
  data = 1:10,
  ncol = 2
)
zeroes_from(A)
}
\references{
\href{https://en.wikipedia.org/wiki/Zero_matrix}{Wikipedia: Zero matrix}
}
\seealso{
Other zeroes functions: 
\code{\link{zeroes}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{zeroes functions}
\keyword{zeroes}
